+++ index.html
@@ -13,7 +13,7 @@
     <title>geOrchestra header</title>
   </head>
   <body style="margin: 0">
-    <geor-header config-file="/sample-config.json" custom-nonce="mynonce" stylesheet="bla.css" height="80"></geor-header>
+    <geor-header  custom-nonce="mynonce" stylesheet="bla.css" height="80"></geor-header>
     <script type="module" src="/src/main.ts" nonce="mynonce"></script>
   </body>
 </html>
