+++ src/ui/GeorchestraLogo.vue
@@ -174,3 +174,4 @@
     </g>
   </svg>
 </template>
+<script setup lang="ts"></script>
