+++ src/ui/MobileMenu.vue
@@ -1,0 +1,90 @@
+<script setup lang="ts">
+import {
+  checkCondition,
+  getItemSelectedTitle,
+  replaceUrlsVariables,
+  state,
+} from '@/shared'
+import type { Dropdown, Link, Separator } from '@/config-interfaces'
+import DropdownItem from '@/ui/DropdownItem.vue'
+import LinkItem from '@/ui/LinkItem.vue'
+import { t } from '@/i18n'
+import ChevronDownIcon from '@/ui/icons/ChevronDownIcon.vue'
+
+const props = defineProps<{
+  menu: (Link | Separator | Dropdown)[]
+}>()
+</script>
+<template>
+  <template v-for="(item, index) in props.menu" :key="index">
+    <template v-if="!item.type && checkCondition(item)">
+      <a
+        :href="(item as Link).url"
+        class="nav-item-mobile"
+        @click="state.activeAppUrl = (item as Link).activeAppUrl"
+        :class="{
+                  active: (item as Link).activeAppUrl == state.activeAppUrl,
+                  disabled: (item as Link).disabled
+                }"
+      >
+        <div class="flex items-center">
+          <span class="ml-1 first-letter:capitalize">
+            {{
+              (item as Link).i18n
+                ? t((item as Link).i18n)
+                : (item as Link).label
+            }}
+          </span>
+        </div>
+      </a>
+    </template>
+    <template v-else-if="item.type === 'dropdown' && checkCondition(item)">
+      <div class="group inline-block relative">
+        <button
+          class="nav-item-mobile after:hover:scale-x-0 flex items-center"
+          @click="toggleDropdown(index)"
+        >
+          <span class="lg:mr-2 md:mr-1 first-letter:capitalize">{{
+            (item as Dropdown).i18n
+              ? t((item as Dropdown).i18n)
+              : (item as Dropdown).label
+          }}</span>
+          <ChevronDownIcon class="w-4 h-4" stroke-width="4"></ChevronDownIcon>
+        </button>
+        <ul
+          class="absolute border rounded w-full dropdown z-[1002] bg-white"
+          v-show="state.activeDropdown === index"
+        >
+          <template
+            v-for="(subitem, subindex) in (item as Dropdown).items"
+            :key="subindex"
+          >
+            <li
+              v-if="checkCondition(subitem)"
+              @click="state.activeAppUrl = (subitem as Link).activeAppUrl"
+              :class="{
+                        active: (subitem as Link).activeAppUrl == state.activeAppUrl,
+                        disabled: (subitem as Link).disabled
+                      }"
+            >
+              <a
+                :href="replaceUrlsVariables(subitem.url)"
+                class="capitalize !flex justify-center items-center"
+              >
+                <i
+                  v-if="subitem.icon"
+                  :class="subitem.icon"
+                  class="pr-1 block pb-[2px] subitem-icon"
+                  style="font-size: 1rem"
+                ></i>
+                <span class="block">{{
+                  subitem.i18n ? t(subitem.i18n) : subitem.label
+                }}</span>
+              </a>
+            </li>
+          </template>
+        </ul>
+      </div>
+    </template>
+  </template>
+</template>
