+++ src/ui/icons/CatalogIcon.vue
