+++ src/ui/icons/ChartPieIcon.vue
