+++ src/ui/icons/MapIcon.vue
