+++ src/ui/icons/UserIcon.vue
@@ -1,12 +1,6 @@
 <template>
   <svg width="1em" height="1em" viewBox="0 0 24 24">
-    <g
-      fill="none"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-    >
+    <g fill="none" stroke="currentColor">
       <path d="M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2"></path>
       <circle cx="9" cy="7" r="4"></circle>
       <path d="m16 11l2 2l4-4"></path>
